added a file to contain location filter related utils
